backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3
watchdog==3.0.0
pillow==10.2.0
requests==2.31.0
